Add persistent-usb.sh script

+++ lfs-scripts/114.sh
@@ -6,4 +6,12 @@ rm -rf /tmp/*
 
 find /usr/lib /usr/libexec -name \*.la -delete
 find /usr -depth -name $(uname -m)-lfs-linux-gnu\* | xargs rm -rf
-userdel -r tester
+
+id tester
+
+if [ $? == '0' ]; then
+   userdel -r tester
+fi
+
+
+
